requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.8
django-cors-headers==3.7.0
django-storages==1.11.1
django-tinymce==3.2.0
export==0.1.2
facebook-sdk==3.1.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
virtualenv==20.2.2
whitenoise==5.2.0
wrapt==1.12.1
